Add routine for clearing capacitor
The capcitor lives on port 0, bit 4 (pin 25). This change configures
the pin as an output and adds a routine for clearing the capacitor,
which must be done after each event. Note that the pin is inverted,
so we write it low, then high to clear, and we have to set it high
initially.